Add button to load code file from server to page
LoadServerFile.js: When rendered for the first time, it ask for API the list of codes and store it on codeFiles, a variavble of context UserCodeFilesContex. On click, its open the SelectCodeModal

SelectCodeModal.js: A modal where the user can select the code file, see the content and load it on code editor

SaveButton.js: Now, after save the new code file, the variable codeFiles, that belongs to context UserCodeFilesContex, will be updated with data of the new file.