Add feature to delete code files
LoadFileButton was modify to now open the componente SelectMutipleCodeModal.js

This component shows a list of buttons, which one represents a code file. On click a button, its show the code on left side. On click on "Load" button,  the modal closes and the code is copy on editor

In this component is also possible to delete many codes at once. There is a checkbox for each file,  just select the ones to remove and click in "Delete"