Correct mix execution on simulation's code between user
Before this commit, when two or more user were in simulation and one of them executed a code, all simulators would execute the code

The problem was the users's containers weren't isolate from each other, so ROS2 messages would go for every user in simulator, because all nodes in containers had the same name

The solution was create a exclusive network to user's container, isolate them. When it's delete, so the network is.