Modify users's container to use host NVIDIA GPU
The script develop_academy.sh now verify the existence of a NVDIA GPU on host machine and create a new variable (GPU_AVAILABLE) on .env. It's receives "1" case true and "0" case false

In docker_utils, the function startUserContainer() uses  GPU_AVAILABLE to create a container that uses host's GPU for gazebo simulation.