Create prototype page to add new exercise
The component CreateExercise.js is a prototype page to add/remove a exercise in system.

In this commit, the  create exercise page receives the name for the new exercises and on click in "Criar",  the exercise is saved in database and is generated the exercise's page.

On delete, the exercise is removed from database and exercises's page files are erased too

For now, the new exercises page will not show a world in gazebo's window, beacause the exercise's creation page don't have a way to insert .launch, .world and .sff files yet