Add on creation exercise a hal.py code generator and send code to be save on server
On creation exercises page, now there are checkboxs to user select what nodes is necessary on HAL code (motors, odometry, etc). On click on button "Gerar", It's made a request to API, especifically to generate_hal() from function_views.py

This functon receives a list with names of all nodes that will be necessary and returns a string with HAL code. To generate the code, it's used the file base_hal.py, a general code to build a specific HAL.py to a exercise.

To each node selected, will be search for his respectivilly .txt file that contains all code necessary to make the node work inside HAL.py.

For example:  its was selected the node "motors", so will be searched for "motors_node.txt", take all command from it and insert on base_hal.py content

With the code returned to creation exercise page, the user can edit it as see fit and then save the exercise. The code will be save on exercise static folder as HAL.py