Add new table "nodes_types", his model and views.
On db.sql, the table "nodes_types" contains unique names of nodes. This table data It's mainly used to the admin select what nodes he would need on a exercise. EX: motors, odometry, laser, etc.

In model.py, was created a new model call NodeType.

In node_type_views.py, was create a view that returns a list with all nodes.

In CreateExercise.js, now is made a request to API to return a list with all nodes in database. Before this commit was used static values.

In hal_views.py, the function generate_hal was update to receive a list of node's ids. Before this commit was a list of names.