Add new table "guide_page_category", update create exercise operations and add new volume on webapp container
"guide_page_category" is a table that contains all categories used on guide pages. Until now this categories wasn't store on database

Now the exercise table have a new column call guide_page_category_id, that contain the id of category it's belongs.

On create exercise page, was a select input where is select the exercise category. Futhermore, there is a new button to insert a teaser image of exercise. The API was updated to permit this actions to happens

About webapp container, now it's have a new volume bind to guide pages local repository. To this be possible, was create a new .env variable called GUIDE_PAGES_ABSOLUTE_PATH, that as the name implies, contains the absolute path of guide page repository