Change in universe_views of function find_by_name() to check_name_availability()
The new function check_name_availability() verify if there is a universe or world with given name. Both models are verify because the universe's name is used as world's name too on exercise creation. A important detail is the search made in both is case insensite.

On exercise_utils, te function createExerciseDatabase was updated, now universe_exists and world_exists made a search of universe and world by name with case insensitive

Was created a new Router called UniverseRouter that contain request functions to API related to universes