Implement use of migration and seed
Before this commit, Database tables and the data were created by sql files. Now is used Django migration to create tables and seeds files to populate the database.

Each seed is a json file, and thanks to functions get_by_natural_key() in models_helper.py, the relation between 2 tables, can be created by using the name. For example in seed Universe, the field "world" receives the name of world used by the universe.

When the webapp container start to execute the seeds, first is search for a filed called last_seed_executed.txt. If exists, the name of last seed executed inside the file is used to garantee only younger seeders are executed. After this, the txt file is updated/created to contain the name of last seeder executed.