Creation of universe step component to be used on exercise multiStep form
On UniverseDataStep, the user can select a list of existing universes to be used on exercise. Is also possible to create a new universe on this step, to do this, must be inform a name and a world file

The component TransferList is responsible for show a list of objects. There is two cards, the card on left contains all itens, while the card on right contains itens selected by the user

The component UploadFileButton only suffer a little correction